HW02 repos done

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepositoryImpl.java
@@ -1,11 +1,16 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.*;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Predicate;
@@ -13,6 +18,8 @@ import java.util.stream.Collectors;
 
 @Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
+    private static final Logger log = LoggerFactory.getLogger(InMemoryMealRepositoryImpl.class);
+
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
@@ -20,6 +27,7 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
         MealsUtil.MEALS.forEach(meal -> this.save(meal, meal.getUserId()));
     }
 
+    //TODO make edit work with checking userID correctly. Now it not saving userID in meal.
     @Override
     public Meal save(Meal meal, int userId) {
         if (meal.isNew()) {
@@ -29,17 +37,23 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
             return meal;
         }
         // treat case: update, but absent in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        if (repository.containsKey(meal.getId()) && Objects.equals(repository.get(meal.getId()).getUserId(), userId))
+            return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        else return null;
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        return Objects.equals(repository.get(id).getUserId(), userId) && repository.remove(id) != null;
+        if (repository.containsKey(id))
+            return Objects.equals(repository.get(id).getUserId(), userId) && repository.remove(id) != null;
+        return false;
     }
 
     @Override
     public Meal get(int id, int userId) {
-        return Objects.equals(repository.get(id).getUserId(), userId) ? repository.get(id) : null;
+        if (repository.containsKey(id))
+            return Objects.equals(repository.get(id).getUserId(), userId) ? repository.get(id) : null;
+        return null;
     }
 
     @Override
